Remove host tag from path in prebuilt/.
There's no need to install multiple OS's versions of the NDK to the
same directory, so we don't need to keep the host tag in the path for
the host tools. This lets us drop find-win-host.cmd, which was just
for figuring out if we were 32-bit or 64-bit Windows.

Change-Id: I38d459ca68425b6ad20cac542cb60b273b0a2f8f