Get LIT working for NDK libc++ tests.
Part of this patch is copying in the libc++ test runner from r244462.
It includes all the work we've done upstream to make testing cross
compiled targets much simpler. The one thing it drops is the use of
`-include nasty_macros.h`, which is just used to make sure we don't
have awful things like _A in our namespace (those are the names used
by some C libraries for the ctype classes).

It then adds all the Android specific testing functionality
(libcxx.android) that is used in the platform. Some day the platform
is actually going to be the upstream for this, but we're going to
suffer through the copy for now (no modifications were needed).

Finally, it adds the NDK specific testing code (libcxx.ndk). This is
thankfully a pretty thin wrapper around the upstream libc++ test
config (albeit with a lot of it disabled/skipped) and the Android
platform test format.

Exception handling doesn't work for ARM because libgcc is getting used
instead of the ARM EHABI unwind library (deja vu). We don't seem to
actually build a separate static library for the ARM unwinder, so
we'll need to fix that separately.

This doesn't link the correct versions of libatomic/libgcc.  We're
never linking thumb, and we're always using arm5.

Current results (run on ~ToT AOSP volantis):

    Expected Passes    : 4322
    Unsupported Tests  : 5
    Unexpected Failures: 418

A lot of those failures are from ODR issues in iostreams (both netd
and the test were trying to run the static constructors for the same
object). This is a separate problem we'll have to solve. That's about
all the diagnosis I've done so far. I'll look closer once this is in
shape and submitted and the noise from ODR has died down.

Bug: http://b/25329002
Change-Id: Ie9858120ac5aa3c872e5cc964e9792df9b3fc8fa