Prevent ODR issues with platform libc++.
Android L and later uses libc++ in platform libraries, and there will
be ODR issues when using libc++ in the NDK on those platforms.

One destructive way this manifests is that the standard streams, which
are initialized by a static constructor, get initialized twice (first
by netd, then by the app), resulting in the app failing to load. This
was resulting in 364 test failures when running on AOSP ToT volantis
(54 remaining afterward).

Change-Id: Ic2af5a570bb0b6677708801fb1aed6cc6f4f15e4