Move ndk-build into the build package.
This finishes cleaning up the various packages so we don't have any
strange overlap or files escaping their directories any more.

When building a monolithic NDK, this will add a shortcut to the new
ndk-build location in the root of the package (to preserve the ability
to run $NDK/ndk-build).

Bug: http://b/25454836
Change-Id: Icb1399be4f2178bcfd1e9160f0db74f0bfe69931