Allow tests to run against NDK in another path.
With this we can run tests from the NDK source path against an NDK in
the out directory. This makes adding new tests or modifying existing
ones easier because you don't have to copy them into the out directory
with each edit.

If NDK is not set in the user's environment, default to the parent of
the test directory.

Change-Id: I5683d4c9f13aea2683361b6f2d9a3471f5676869