Update build files for GCC update.
The main fix here is removing the prebuilt/ directory from the
toolchain packages. The layout of the GCC package has changed such
that the directory is no longer there (to facilitate using the same
package for the platform), and it's easier to remove it from the NDK
than to put it back (and it isn't needed anyway).

This also update the LLVM package to match.

Finally, this changes the GCC versioned path from 4.9.x-google to just
4.9 which is the fix for http://b/25487043.

Change-Id: Id556b97f42222dafc2ce7a0034b1c7a3606a28e5