Remove test for x86 gcc stack protector hack.
We used to have a hack in GCC to protect against the referenced issue
(http://b.android.com/38121), but we have since removed that
(https://android-review.googlesource.com/#/c/182731/).

This test was really checking the behavior of the compiler, not libc
(see the -static). We already test that we don't have this problem as
part of the bionic tests, so it adds no value here.

Change-Id: If59fb4d28475995473953ac915b1e60254456164