Make one platform package per API level.
The NDK is getting huge, and the reason is that the platforms/
directory is growing without bound. We're making the assumption that
most users don't care about having *every* platform installed, since
they probably are only targeting one, maybe two.

This split is the first step of making it possible to skip installing
some.

Bug: http://b/25954102
Change-Id: Ib5c5fbbceffb30942b9c51ba7cd28cdb10abe13b