Switch all modules from zipfiles to tarballs.
The SDK manager doesn't speak tar. This should also speed up the
package validation step of package.py (since it isn't reading the
files but is just interested in the TOC). It will, however, increase
the size of each module. We're still building the final monolithic
package as a tarball for Linux and Darwin though, so the largest
download will still be the same size.

Bug: http://b/26237669
Change-Id: I967d0b60f4132c864726c2e88f0f52a012b03909