Differentiate OUT_DIR and DIST_DIR.
DIST_DIR is archived on the build server, so we don't want to be
putting everything in there.

Also, this lays the groundwork for us to remove the build -> pack ->
unpack -> repack -> unpack -> test workflow that we currently have
and replace it with build -> test -> pack, since we can install
things to their proper locations, test them where they are, and then
package it directly.

Change-Id: I37529b5f564d82ba0e8b126d9fd44b4b148c1478