Add support for NDK_TOOLCHAINS_ROOT.
This mirrors NDK_PLATFORMS_ROOT, which allows the user to point
ndk-build at a platforms tree that is not in the normal location
relative to ndk-build.

This is particularly helpful if we want to use prebuilt toolchains for
building an NDK module. Specifically, we will use this to build libc++
using ndk-build instead of the old build-cxx-stl.sh and maintaining
two sets of build rules.

Change-Id: I137b1239ebc8068ab63e2fdaa95aff68466e326b