Build libc++ with ndk-build instead of bash.
Before this patch we had to maintain two sets of build rules for
libc++: build-cxx-stl.sh and Android.mk. This changes the prebuild
step from using the shell script to just invoking ndk-build.

For one, this means that anyone using LIBCXX_FORCE_REBUILD will
actually get a reliable result, as it's the same way we built it.

It's also quite a bit faster. It now takes only one minute to build
all 11 ABIs (32-bit ARM ABIs counted twice because we build both
thumb and non-thumb), whereas before this patch it took 3 minutes and
15 seconds.  Aside from the cost of building everything, we also
benefit from dependency checking now since we store the intermediate
files in a reliable location, which means that a null build doesn't
take long (it still takes some time because we have to run the
packaging step).

Now that this has been done for libc++, we could also do this for
stlport. It may not be worth the effort, however, since that's not
being actively worked on.

WIP: I still need to run through build-cxx-stl.sh to make sure we are
building things the same way we used to, which I seriously doubt. I
also need to run this through the test suites still...

Change-Id: I8602f0fd3a23a2938cdc5be19d0714d5567ce6fc