WIP: Add a script to automate more testing.
The problem with run-all.py is that it still requires you to set up
devices and configurations yourself. This script is a wrapper around
run-all.py that will inspect all connected devices, find the ones
that fit a testing configuration (defined in source for now), and
then runs the tests over the fleet.

WIP bits:

Reporting could use some work. Should probably teach run-all.py a new
output format so we can print the useful bits here. The existing
reporting would also benefit from some color and better formatting.

I'm not sure if this should be a separate script, or if I should be
massaging these features into run-all.py.

Logs currently dump into the CWD, which is annoying for git. I should
probably move them to an out dir if we have one or temp as a
fallback. Or maybe I should just leave it as is because there's a
--log-dir option.

There's a lot of copy paste of the desired test configurations.
Should consolidate that so we only have to update one place.

Change-Id: I4f702409e059f8e7d70ca7ab288b6bb9d6dc5b9a