Add a script to automate more testing.
The problem with run-all.py is that it still requires you to set up
devices and configurations yourself. This script is a wrapper around
run-all.py that will inspect all connected devices, find the ones
that fit a testing configuration (defined in source for now), and
then runs the tests over the fleet.

There are still some improvements to be made here:
 * Still some duplication for slight behavioral differences that
   should be shared.
 * Output readability could be improved.
 * Still might want to fold this into run-all.py as --discover or
   something.

But this is already very useful as is, so I think it's time to submit.

Change-Id: I4f702409e059f8e7d70ca7ab288b6bb9d6dc5b9a
(cherry picked from commit 65555b0ba4179357f1d7e23b5ae4d40224c005cd)