Differentiate build number and release name.
For a real release we want --release to be "r11" or similar, but we
still want the build number data. Split this into two options so we
can do both, and wire up all the defaults.

With `--release r11 --build-number 42`, the package will be named
"android-ndk-r11" and source.properties will use the version
"11.0.42".

With `--build-number 42` the package will be named "android-ndk-42"
and source.properties will use the version "42.0.42".

With neither argument, the package will be named "android-ndk-dev"
and source.properties will use the version "dev.0.0".

Using --release without --build-number is untested (can't think of a
good reason to do that).

Change-Id: Id7cd980b9af5ac3fd44ce7894ee99f33a212741c