Refactor test running code.
We want to be able to reuse most of this code for to workflow where
the test runner uses all the devices it can find as well at the case
where we're running against a user defined device and configuration.
Separate the code that runs the tests themselves from the code that
handles the command line and initial setup.

A huge amount of this code is going to need to be moved into another
file so it can actually be imported (I shouldn't have put a hyphen in
a Python file name), but I'm doing that in a follow up patch to
hopefully make this review easier.

Change-Id: I767f7fd13747b68e6221386b3dd867a8c8f29047