Print test results immediately.
Previously the test runner would wait until all the tests had been
run and print the results during the summary. Making failures visible
as early as possible is better. This patch fixes that.

Since we try to keep the concern of whether a test was expected to
fail or not separate from actually running the test (which we
currently do for build tests, but don't for device tests), this means
the printing of the test status as we go needs to come up into the
test runner where we perform the xfail fixups. To do so without
having to wait for every subtest to finish before printing, Test.run
now returns a generator instead of a list.

This also pushes the reponsibility of actually printing each result
down into the Printer.

Change-Id: I1a09a9a3402da1bca3751ff7a3e3583a52620326
(cherry picked from commit 7a6977a38ada0bc358ddb0ce022123e6d8c4a50c)