Remove useless hard coded max platform level.
We already have a block down below that does this in a much better
way (using an obviously named constant from init.mk that is
initialized by scanning the platforms/ directory for the latetst
available platform).

IMO setting `APP_PLATFORM := android-42` and having that be a moving
target up until we actually release android-42 is pretty surprising
behavior. Might be better to allow "android-latest" and explicitly
fail if we're given a version we don't know about.

Change-Id: I31dca24dadab46365fb9949061bca8325ce1bab6