Added a testcase for NEON
A simple test case to validate that NEON code can be compiled and
launched for all supported targets and toolchains.

Also added clang support for neon2sse compatibility layer to enable NEON
on x86 and x86_64 devices. If an application uses clang and neon and is
built for x86*, then the corresponding GCC header is used.

Change-Id: I8c3127b4c2b949997595aaf9ce4f97e659a8e0c5
Signed-off-by: Egor Kochetov <egor.kochetov@intel.com>