Ensure that the unwinder is linked hidden.
The unwinder needs to be linked hidden statically into each binary to
make sure only one unwinder will ever be used for any given call site.

Using `-fvisibility=hidden` is insufficient because some things in
the unwinder are marked with default visibility, and we can't just
fix the `_LIBUNWIND_EXPORT` definition either because most of the
headers just use `extern`. We should fix this upstream, but we're
already way out of date so this is the short term fix.

Change-Id: Ib6c82cea423ff843009f3eb2ad5cd30722a80645