Build the device tests during checkbuild too.
Right now we have the build server running the build tests, but we
should check that the device tests are at least buildable even if we
don't have a device to run them on.

This adds support for skipping the execution phase of the tests to
`run_single_configuration`. This is exposed from `run-all.py` as
`--skip-tests`.

Change-Id: Ie129041bc3104ad50b3e96c23158d6918d36b25a