Correct a couple test configs.
Using `match_broken` on a device test without checking the subtest
means that the test *build* is broken. Right now there's no visible
difference for tests that contain only a single test case, but I'm
about to add support for building the device tests even if we don't
run them, at which point we will need to get this right.

Change-Id: I1cc2ceadccf4844b0f392510a7a7cd13f8115416