Parallelize the build.
This shows a roughly 2x speed up of the build step (ignoring tests
and packaging).

Both times run with:

    $ time ./checkbuild.py --arch arm64 --no-test --no-package

The serial result:

    real    16m19.286s
    user    19m44.367s
    sys     6m2.884s

The parallel result:

    real    7m44.093s
    user    26m16.921s
    sys     8m1.321s

The only problem is that SIGINT will spew a lot of garbage to the
terminal. From what I could figure out it looks like this is actually
just something to do with the multiprocessing pools. Might dig more
later, but it seems worth getting this in even with that.

Bug: http://b/28214235
Change-Id: Iff48f31f49255f5482c36fd86b1fc1ac0faa4958