Make to the support requirement more obvious.
I don't know if people aren't reading or just don't care, but we've had
a greater than usual number of bugs filed against unsupported NDKs
recently. For the people that just didn't see the previous text about
not filing bugs for unsupported NDKs, maybe this will help.

Some of those are filing against "canary" afaict just because they're
not using any of the supported versions, but aren't using a canary
either. I can't remember the last time anyone's used the "canary" input
correctly, so just remove that option.

I also added a required checkbox for "I'm using a supported NDK", with
an explicit "don't file the bug if it's not a supported NDK"
description.